+++ CHANGELOG.md
@@ -14,6 +14,51 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2024-XX-XX
 
+## [v5.7.0] 2024-10-08
+
+- [add] Access control: Viewing rights.
+
+  When a user's viewing rights have been revoked:
+
+  - SearchPage redirects to NoAccessPage
+  - ListingPage redirects to NoAccessPage unless the user is viewing their own listing. Reviews are
+    not displayed.
+  - ProfilePage redirects to NoAccessPage unless the user is viewing their own profile. Reviews are
+    not displayed.
+  - TransactionPage does not show the OrderPanel for transactions at the inquiry state of booking or
+    purchase processes.
+
+  [#459](https://github.com/sharetribe/web-template/pull/459)
+
+  [v5.7.0]: https://github.com/sharetribe/web-template/compare/v5.6.1...v5.7.0
+
+## [v5.6.1] 2024-10-07
+
+- [fix] Updated JSON asset schema structure from Console: Removed the enabled field, now using only
+  the type attribute for CTA status.
+- [add] Add currently available translations for DE, ES, FR.
+  [#461](https://github.com/sharetribe/web-template/pull/461)
+
+  [v5.6.1]: https://github.com/sharetribe/web-template/compare/v5.6.0...v5.6.1
+
+## [v5.6.0] 2024-10-01
+
+- [add] Add support for CTAs on NoAccessPage
+  [#455](https://github.com/sharetribe/web-template/pull/455)
+- [add] Add currently available translations for DE, ES, FR.
+  [#458](https://github.com/sharetribe/web-template/pull/458)
+- [fix] Topbar: malformed custom link causes a 500 error on server.
+  [#457](https://github.com/sharetribe/web-template/pull/457)
+- [fix] ListingPage.shared.js: import for convertMoneyToNumber was not made for priceForSchemaMaybe
+  function. [#456](https://github.com/sharetribe/web-template/pull/456)
+- [add] Access control: Transaction rights
+
+  - When a user does not have the "initiateTransactions" permission in their
+    `effectivePermissionSet` relationship and they try to initiate an order or send an inquiry, they
+    are redirected to the NoAccessPage.
+
+  [#450](https://github.com/sharetribe/web-template/pull/450)
+
 - [fix] InquiryForm: test code was committed earlier.
   [#452](https://github.com/sharetribe/web-template/pull/452)
 - [fix] EmailVerification: enforce that currentUser is fetched after verification.
@@ -24,6 +69,8 @@ way to update this template, but currently, we follow a pattern:
   [#448](https://github.com/sharetribe/web-template/pull/448)
 - [fix] Currencies that the Stripe does not support should not cause 500 errors.
   [#447](https://github.com/sharetribe/web-template/pull/447)
+
+  [v5.6.0]: https://github.com/sharetribe/web-template/compare/v5.5.0...v5.6.0
 
 ## [v5.5.0] 2024-09-03